Fix: Resolve Stripe Elements context error
The `useStripe()` and `useElements()` hooks in `DonationForm.tsx` were being called outside of an `<Elements>` provider. This was because the `<Elements>` provider was incorrectly located within the `StripeElementsWrapper.tsx` component, which was a child of `DonationForm.tsx` rather than an ancestor.

I refactored the Stripe Elements setup:
1. I moved the `<Elements>` provider from `StripeElementsWrapper.tsx` to `app/(pages)/donate/page.tsx`, so it correctly wraps the `DonationForm` component.
2. I renamed `StripeElementsWrapper.tsx` to `StripeCardInput.tsx` and it now solely renders the `CardElement` and its associated label, without managing the Elements context.
3. I updated `DonationForm.tsx` to use the new `StripeCardInput.tsx`.

These changes ensure that `useStripe()` and `useElements()` are called within the scope of an `<Elements>` provider, resolving the "Could not find Elements context" error.
