+++ app/(pages)/donate/page.tsx
@@ -1,9 +1,15 @@
 // app/(pages)/donate/page.tsx
 import React from 'react';
-import DonationForm from '@/components/DonationForm';
-import Cover from '@/components/Cover'; // Optional: for a consistent page header
-// You might want to add a specific cover image for the donation page
-import coverImage from '@/public/images/herocover.png'; // Placeholder, replace with a relevant image
+import DonationForm from '@/components/donation/DonationForm'; // Corrected path
+import Cover from '@/components/Cover';
+import coverImage from '@/public/images/herocover.png';
+
+// Import Stripe libraries
+import { loadStripe } from '@stripe/stripe-js';
+import { Elements } from '@stripe/react-stripe-js';
+
+// Load your publishable key from environment variables
+const stripePromise = loadStripe(process.env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY || 'your_stripe_publishable_key_here_backup'); // Added a backup key just in case
 
 const DonatePage: React.FC = () => {
   return (
@@ -17,12 +23,10 @@ const DonatePage: React.FC = () => {
           <p className="text-lg text-gray-700 mb-8 text-center">
             Your generous contribution helps us empower Zomi youths worldwide. Thank you for your support!
           </p>
-          {/* 
-            The Stripe Elements provider should be an ancestor of any component using useStripe or useElements.
-            In this setup, DonationForm imports StripeElementsWrapper, which contains the <Elements> provider.
-            So, we can directly use DonationForm here.
-          */}
-          <DonationForm />
+          {/* Wrap DonationForm with Elements provider */}
+          <Elements stripe={stripePromise}>
+            <DonationForm />
+          </Elements>
         </div>
       </main>
     </div>
